requirements.txt
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.7.2
seaborn>=0.12.2
beautifulsoup4>=4.12.2
wordcloud>=1.9.3
python-dateutil>=2.8.2
PyYAML>=6.0